Refactor DataTable utility imports and remove unused deep-utils file
Update import paths for utility functions in DataTable components to streamline code organization. Remove the now redundant deep-utils file, which contained utility functions that are no longer in use, enhancing maintainability and clarity of the codebase.
